requirements.txt
django >= 3.2
django-dotenv
djangorestframework
django-cors-headers
requests
gunicorn
psycopg2-binary 
